move to `src`, add space, add space in end
